Passing serialize and parse byte tests.
Now passing the serialize and parse byte tests, a single test each is
passing for both the serializer and the parser.

One test was converted to node for the sake of debugging because it
triggered an endless loop that prevented `npm test` from completing.
That test still does not past, but it does not loop forever.