Fixed failure of parsing of bit packed patterns.
The bit packing pattern was slurping the entire pattern, not terminating
at the ending curly brace. It was using a dot to match the contents of
the bit packing pattern. I've replaced the dot matching with a pattern
that will terminate at the closing curly brance and allow for a single
nested level of curly braces in order to specify fill bytes for skips.

Closes #19.
Closes #7.