Serialize alternate structures.
Implemented serialization of alternate structures, that is, alterate
patterns that define multiple fields. Currently, a tangled mess in
there, so I've opened #34 to come back and convert that to a queue or
stack driven process at a latter date.

Update the `README.md` to indicate that bit packed and alternate
patterns do work, plus added documentation on alternate structures.

Closes #31.