API purge.
Removed all stream related stuff from the API. Those concepts will be
moved to their own project to be called upon when needed. They are not
always applicable.

Moving forward, we remove the range checking from the Serializer,
instead counting on the user to use `sizeOf` to verify the size of the
`Buffer`, if there is a chance that the buffer will overrun. Many
applications will have fixed size packets, so it will be a waste of
effort on the part of `Serializer` to test for buffer overrun.

Note that buffer overrun in benign in Node.js. It is not the same sort
security threat that you have in C.

All of the code except for our IEEE754 library and the pattern parser
has been moved into the top level `index.js`.