Convert to closure style.
Not a complete conversion, but a good stopping point. Removed the
`Packet` base class, using a member class named `Definition` instead.
Most of the member variables have been converted into variables enclosed
in closures. The constructors for `Parser` and `Serializer` are no
longer exposed. Instances of the objects are instead created thorugh
factory functions, continuing the radical change to the API.

Rewriting the `proof.js` harnesses was enough to get most of the tests
to pass with the new API. Few of the tests needed to be rewritten since
they do not use the API directly, but instead use helper functions.

Closes #40.
See #39.
See #38.