Add `.js` suffix to tests.
I've chosen Perl conventions for test directory layout and test file
names, so now GitHub is marking my JavaScript repositories as Perl
repositories. The help article [My repo is marked as the wrong
language](https://help.github.com/articles/my-repo-is-marked-as-the-wrong-language)
sends you off to look at [Linguist](https://github.com/github/linguist).

The Linguist library will mark the source as Perl. Then, when you remove
Perl from the YAML file that defines languages, it will mark the library
as Turing. Perhaps the bayesian classifier will prefer Perl for my
source files, or perhaps it has some sort of hard coded override to
prefer Perl.

When I copy the Packet index.js to index.t and take it out of t
directory, Linguist will still consider it to be Perl, not Turing nor
JavaScript.

My choices are to either I submit a patch to Linguist or else I rename
all my tests to *.t.js. The latter is the shortest path to getting my
files listed correctly. Any patch would need to be accompanied by
justifications that are likely to be rejected.

Removing the file extension also works, but I want to have something to
glob into in the test directories so that there can be proof.js files in
the same directory.