Copy pattern after to alternation.
After determining the branch of an alternation and before updating the
pattern to replace the alternation pattern with the selected branch,
create a copy of the working pattern.

If the working pattern was a named pattern read from the set of
pre-defined named patterns, the working pattern will reference the
pattern in the pattern cache. Updating the pattern will update the
pattern in the pattern cache. This gives the impression that alternation
is always selecting the same same branch, when in fact the cached
pattern has been rewritten with the selected branch, so the alternation
has been replaced by the pattern parts in the branch chosen on the first
run of the named pattern.