Stop serialization using index.
Stopping serialization by testing if the pattern index is equal to the
pattern length instead of setting the pattern to null. When we implement
`sizeOf` we're not going to want to report the size of the write even
after the buffer has been written, so we don't want to lose the pattern.