Remove asterisk to indicate wildcard condition.
Removed the asterisk to indicate a wildcard alternate. Any alternate
without a condition is a matches all conditions, any alternations after
are not going to be considered.

When removing the asterisk, I encountered tests where the asterisk was
being used with a split condition, so I thought I'd made a mistake in
calling for its removal. However, an alternate maps to a binary
structure. That binary structure is going to have a single
representation, so I don't see how a default for the read pattern is
going to be different from the default for the write pattern.

If it is different somehow, if there is some complexity I'm not
considering, then the user can still write separate read and write
alternations.

Closes #33.