Add `offset` argument to `offsetsOf`.
Added an optional buffer start. Not going to require it because the
`buffer` is not required, but the `offset` is necessary to calculate
offsets in the absence of a buffer. Thus, the signature would have an
optional `buffer` but a required `offset`. I don't want to document
that. `offsetsOf` is already as performant as it is going to be. Unlike
the composition strategy, the function gets initialized once. It is used
for pattern and packet inspection, visualization, so it is not as
performance critical as `write`.

Closes #101.