Covert parser tests to use `Proof` `counter`.
Use the `counter` function in `Proof` to set the test count for each
invocation of `parseEqual`. This will allow us to add tests to
`parseEqual` without having to go back and update the test count of each
test.

There's a new option in the `parseEqual` options array, `count`. If it
is defined, this is the count passes to `counter`. There is one test
that invokes `parseEqual` counting upon a failure.

Closes #200.