`parse` test function is incremental by default.
The `parse` test function is incremental by default. If you do not
provide a specific parition for the buffer reads, then after a full
buffer read, `parse` will parse the buffer once for each offset. It will
reset the parser, then read the buffer from zero to the offset, then
read a byte at a time thereafter, once for each offset. This will
exercise all the fallbacks and all the incremental branches.

Some tests broke, but rather than try to fix the exsiting generic
parser, which is outgoing, I moved them out of the tests with a rename.

I added a report to `t/test` that lists all the broken tests it finds.
It will remind me to turn these tests on as I implement the new compiled
parsers and serializers.

I'm going to tidy as go with the tests. This commit tides
`t/parser/words.t.js`.

Closes #214.
See #218.