Serialize test resets buffer, fix serializer skip.
The serializer test did not reset the buffer. The incremental serializer tests
reuse the same buffer. If the incremental serializer is broken in such a way
that it does not write to the buffer, then the test passes based on what was
written by the non-incremental serializer.

Fixing this reveals that the incremental serializer skip was broken, not
tracking where it left off in the buffer.

Closes #221.
Closes #220.