Remove subsequent serializer detection.
I'll add it later when I recreate the API. It's not used now. The
current API and class architecture is really just an extension of the
test harness for the compiler. It's all outgoing.

In the mean time, I'd like to see 100% test coverage on the new parsers
and serializers as I create them.

See #175.