Rename generated files.
I'd been naming them serialier and parser, but they are functional and
not object oriented. Even in their interactions with the object
orientation, even then they are a function on an object. The new
compilers use the verbs, parse and serialize, instead of the nouns
parser and serializer. As I create the new compilers, I'm going to
delete files and add new ones, when I'd rather see the file get
completely rewritten.

While I'm renaming things, I believe I'm not testing parsers and
serializers, but parsing and serialization, so it's time to rename the
test suites once again, from parser and serializer, to parse and
serialize.

A note about behavior. I sweat these renames when they occur, because I
don't want to have a horrible `git` history, but I make them none the
less. If at some point, I feel that I'm working with machines and not
operations, I'll probably agonize over whether using nouns in the test
names is better than verbs, and agonize over the change because it is
going to create such a huge change set.

Given that these are tests, and not the public facing API, I'm not
agonizing over an adverse effect on the user. It's agony wasted.

Closes #231.