Specify subsequent pattern from callback.
Resurrection of the internal state machine, which is what it is, for
both the parsers and serializers. I've tested the best-foot forward part
of the best-foot forward parser and serializers using the cycle test in
this commit. I'll test the incremental parser in a subsequent commit.

Becoming concerned about the whitespace. I don't like the two blank
lines before the callback in the best-foot forward section of the
parsers.

See #241.
See #175.