Pretty print double quotes.
Currently, the source generated contains double quotes for object keys
instead of single quotes, which is the style of the code. It's easier to
change the pretty printer for now than to change the existing template
based code generation. I'd like to have parity for a time, the rewrite
generates the exact same parser.

Closes #274.