Compiler can no longer specify new lines.
Compiler can no longer specify new lines itself. It was detecting a
lonely new line and converting it to the indicator for the `prettify`
function; "__nl__". This is interfering with the desire to strip new
lines. Obviously, we need to be explicit about new lines when we invoke
the compiler, specifying our new lines in our generated source.

Closes #285.