Use existing generated parsers.
In order to begin to test the new concatenation generated parser
without having to re-implement the cycle test, or partially disable it,
I've updated the interface to use the existing generated parsers,
serializers and sizers that were generated by the template based
generators.

Also, extracted the name generation of a generator module to it's own
module.

Closes #290.