Implement fixed arrays.
Updated Programmatic to allow for dedenting after interpolating code
that is itself indented. It requires that I used the `// __reference__`
hack to dedent the interpolated code. Ideally, Programmatic would dedent
all on it's own. I'll have to add a test and an todo to Programmatic.

Looping over an array has now been implemented. I'm not sure how far
nested this looping can go, though. How does one have an array of
zero-terminated key/value strings? Would it require a significant change
to the specification language where curly braces become blocks?

Considering how to make the functions generated simpler. Currently I
pass in a callback so that when the parser is finished, it can invoke
the user's callback. The function is invoked with a `this` object so
that if the inline parser cannot finish parsing, it can change the
`parse` method of the `this` object.

Ideally, I'd return `null` if the inline parse completed, or the
incremental parser to use if it didn't.

However, I also need to return the new start position; the number of
bytes consumed.

I'm almost back to where I began, wondering why I don't just assign
properties to an object, but I do not not like the idea of having the
generated parsers invoke the callbacks.

I'd like to break it out so I could experiment with efficiency of
different methods, but I'm going to have to assume that it is more
efficient for an outer loop to invoke user callbacks.

Thus, I see that there are three bits of information returned from a
parse; the updated start position, the next function to use if any, and
whether or not we are done. The object construction itself is by
reference and passing, seemingly like a side effect.

I've used a `null` function to indicate that we are done.

I'm now leaning toward passing in arrays as references, or possibly a
single array with the different parts.