Explode language for incremental parsers.
Implement incremental parsers using an exploded language, a langauge
that is more specific, more verbose than the current terse language that
aims to be human readable. This machine readable AST will make it easier
to develop the generators. A translator between the human readable
language and the machine readable language will be a separate component
within this project.