Incremental structures always created in switch.
Structures created in an incremental parser, including the root
structure, are always created in the switch statement. The root element
of the stack is a dummy node. The top most object on the stack is always
named `object` whereas the root will probably pull its name from the
language definition.