Sketch of language, intermediate format generation.
Creating a new syntax bashing only language. No real unit tests, just
printing to standard out. Committing now because it is looking nice, a
lot of work is done, and I'm liable to break things soon, will want to
see a diff when I do.

Once again, I came across the concept of compressed or encoded integers.
Looks as though they are their own concept that need special
accommodation, but do not require the extension of the language to
support generic looping. They will be a fixed length based on
information in the stream, they are not going to be arbitrarily
calculated and if they are then Packet can't help you, just as a tool
like ANTLR can't parse Perl.

Compressed integers in the case of MQTT present a problem because a
compressed integer is used in the header to represent the size of the
remaining message. You won't know the length of the size field until the
reset of the packet is serialized. This is the always the way, though,
with length in header or worse, checksum in header because you can't
even do a size of calculation, you have to actually serialize, but its
made even worse by the fact that you don't know how many bytes to skip.

So, parsing can be incremental, but serialization will need look-ahead
options at times. We can attempt to serialize to buffer, but if we run
out, we have to serialize to an array to get our checksum and length. We
can say that this is price you pay for falling back to incremental.
Otherwise, we leave the minimal amount of space for our compressed
integer, and if it requires more space than specified we have to move
the memory forward. We pay this one regardless of whether the buffer can
accommodate the packet whenever the length is larger than the minimum.

Our intermediate output is going to need the min and max length of any
conditional that sets a calculated value.

Provisions for checksums where a blocker in the earlier iteration of
this library. No good notion of how to express them in the language at
this juncture. Haven't thought about it too hard, but there seems to be
plenty of room left in the language for such complexities.

Upshot is that if you're designing a format and you have a checksum, put
it in a footer. If you're going to encode the length, create a chunked
encoding with a headers for each chunk if you want to save header bytes.
You'd pay for long messages by repeating the chunk header only if
they messages are indeed long.

See #400.