Block scope all, path index inc.
Block scope variables make all the difference for the whole packet
parser. There will be a rule that you cannot reuse a property name on a
particular path (i.e. no `object.header.object`) so that we can simply
assign the property to a block scope variable of the same name. In
homage to Perl we assign the immediate actual value to `$_`, but the
user can just assume that for `object.header` there is an `object` and a
`header` in scope when they are writing injected JavaScript code.

Now there is no more tracking varibales for hoisting, they are declared
with `let` as needed. I've gotten length-encoded serialization of
objects with integers working. Now I need to stop and create a cycle
framework so I can do each feature in mind and have it done, with a
target of covering the features in the historic readme and parsing MQTT.

A stretch goal might be to parse `tar` which was the first project I
undertook and it was too much for the cute string-of-mnemonics language.

The incremental parser will avoid collisions in its namespace flattened
by the switch statement used for control-flow by always obtaining the
immediate actual value from the full path to the value in the root
object. I'm tracking indices into arrays in an array of indicies and
from the whole-packet parser in a best-foot-forward parser I pass in an
array literal constructed from the local index values in scope. You have
to see it, really. Go look at the generated best-foot-forward length
encoded serializer and incremental seralizer named `nested` here.

Will be breaking things out into much smaller steps.

Closes #194.
Closes #134.