Track depth of index in index array.
Seems that for now we can use an index array, but we might want to use
index variables and build that array when we fallback to incremental. An
alternative naming convention would be to prefix the index with a dollar
sign or two dollar signs in case someone names an array `buffer`.

See #258.