Create an object to track accumulation.
Since we use the same function to generate accumulators and inlines in
the different generators, we should have a common object that gathers
the accumulator/inline specific properties to simplify these function
signatures. Don't want to create an object, but maybe this is better as
a closure of some sort?

See #376.