Serialize and parse fixed byte arrays as buffers.
Implement serialization and parse of buffers, fixed and terminated.
Incremental parsing padding fixed arrays is not redirected to the
terminated parser implementation which will add checks for reaching the
maxiumu fixed length array. Incremental terminated buffer parsing is
implemented, but not serialization and not synchrnous operation.

Buffers are either catenated, meaning during incremental parse, multiple
buffers are gathered, than catenated into a single buffer, or else they
are left as an array of buffers which are themselves slices of the
underlying buffer.

Closes #506.
Closes #513.
Closes #517.