Generate a checkpoint at every field for testing.
Check that we are aligning the steps between best-foot-forward and
incremental parsers correctly by generating a special best-foot-forward
serializers and parsers that are liberal with the application of
checkpoints. The inquisition serializers and parsers will place a
checkpoint before each field instead of rolling up a checkpoint for a
series of fixed fields. This makes it easier to test that a field
best-foot-forward implementation is correctly tracking the steps in the
incremental implementation. Placing a single byte after the tested field
will reveal of the set for the checkpoint before that byte is correctly
offset.

Closes #531.