Only invoke accumulators as necessary in packed.
For packed conditionals that may referenced accumulators, we know that
any of the conditionals are at the same byte position in the underlying
buffer so if two conditionals test reference the same accumulator we do
not need to run the accumulator once for each test, just once for the
serialized packed integer.

See #563.