Implement inferred spread integers.
If a spread integer definition is shorter than the byte length of the
integer, fill the definition by duplicating the second element. This
will fill the middle of the definition, which accommodates the common
case of a leading byte followed by a repeat of subsequent bytes, or a
leading a trailing byte that is different from the middle bytes.

Had to group upper bit mask definition and byte size in an array when
upper bit mask is defined to get a consistent length of the definition
part of the array. Apparently there was room in the language for this.

Closes #589.