Run includes prior to interpretation.
Including partials is pretty explicit. We replace any string that maps
to a snippet with the snippet. We can do this in an separate pass and
simplify `language.js`. The include function is added here and now, and
in a separate commit we are going to remove any special handling to
disambiguate snippets from `language.js`.

See #610.