Fix inlines around single bit packed field.
Single bit or any bit where modulo 8 returns 1. The language parser was
not propagating the `pack` flag to the integer function when it
descended into the wrapped field in an inline. If there is a remainder
of `1` for that value it triggers endianess interpretation. The number
value gets a bitwise not which turns it to `0`.

Closes #612.