Add elision test with no elision.
Put a test in place for accumulators, but ran into an issue implementing
the elision.

Starting to see how accumlators may or may not be elided, but so too
should a definition that begins with an inline. We need to mark the
elision before we descend the tree, not handle it as a special case at
the property. In a general sense. Probably have to update the AST.
Definately have to update the vivifier.

See #588.