Rewrite `vivify.structure` for elision.
Elision needs to descend from the definition of the field property name
in the structure definition down to where the actual field is defined
passing through inlines and accumulators. We also have to track names as
we descend because the field name is at the root of the declaration of
the field in the AST where it may be an inline or an accumulator. We
should find a way to move that down or at least have the structure of
the AST tree be defined and understood so that this descent of the name
is no longer described as "at least this works."

Would like to know how many objects are 'decent' vivified versus leaf
vivified. It is not many and it should not take long to sort this out.

See #588.